Use unauthenticated encryption for auth challenges
Auth challenges using authenticated encryption (AES-GCM) was a major
security flaw, allowing a user's password to be guessed an infinite
number of times with a single challenge request, subverting
ratelimiting. This has been fixed by using AES-CTR encryption instead of
GCM to encrypt challenges, meaning that clients will be forced to rely
on the server (and its ratelimiting) to verify a password attempt.
Chosen-ciphertext attacks are still prevented by using authenticated
encryption for TLS communication with the API.